+++ src/main/java/com/robertx22/age_of_exile/database/registry/RegistryPackets.java
@@ -3,8 +3,8 @@ package com.robertx22.age_of_exile.database.registry;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSyntaxException;
 import com.robertx22.age_of_exile.aoe_data.datapacks.bases.ISerializedRegistryEntry;
-import com.robertx22.age_of_exile.auto_comp.ItemAutoPowerLevels;
 import com.robertx22.age_of_exile.database.IByteBuf;
+import com.robertx22.age_of_exile.uncommon.auto_comp.ItemAutoPowerLevels;
 import com.robertx22.age_of_exile.uncommon.testing.Watch;
 
 import java.util.ArrayList;